@@ -13,10 +13,10 @@
     <name>dummy-service</name>
     <description>Demo project for Spring Boot</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <frontend-src-dir>${project.basedir}/ui</frontend-src-dir>
-        <node.version>v14.14.0</node.version>
-        <yarn.version>v1.22.5</yarn.version>
+        <node.version>v23.1.0</node.version>
+        <yarn.version>v1.22.22</yarn.version>
         <frontend-maven-plugin.version>1.9.1</frontend-maven-plugin.version>
         <sonar.projectKey>davidesalerno_dummy-service</sonar.projectKey>
         <sonar.organization>davidesalerno</sonar.organization>
